Move centralized swap plugins to a folder
