+++ src/index.ts
@@ -2,8 +2,13 @@ import 'regenerator-runtime/runtime'
 
 import type { EdgeCorePlugins } from 'edge-core-js/types'
 
-import { makeChangeHeroPlugin } from './swap/changehero'
-import { makeChangeNowPlugin } from './swap/changenow'
+import { makeChangeHeroPlugin } from './swap/central/changehero'
+import { makeChangeNowPlugin } from './swap/central/changenow'
+import { makeExolixPlugin } from './swap/central/exolix'
+import { makeGodexPlugin } from './swap/central/godex'
+import { makeLetsExchangePlugin } from './swap/central/letsexchange'
+import { makeSideshiftPlugin } from './swap/central/sideshift'
+import { makeSwapuzPlugin } from './swap/central/swapuz'
 import { makeLifiPlugin } from './swap/defi/lifi'
 import { makeThorchainPlugin } from './swap/defi/thorchain'
 import { makeThorchainDaPlugin } from './swap/defi/thorchainDa'
@@ -11,11 +16,6 @@ import { makeSpookySwapPlugin } from './swap/defi/uni-v2-based/plugins/spookySwa
 import { makeTombSwapPlugin } from './swap/defi/uni-v2-based/plugins/tombSwap'
 import { makeVelodromePlugin } from './swap/defi/uni-v2-based/plugins/velodrome'
 import { makeXrpDexPlugin } from './swap/defi/xrpDex'
-import { makeExolixPlugin } from './swap/exolix'
-import { makeGodexPlugin } from './swap/godex'
-import { makeLetsExchangePlugin } from './swap/letsexchange'
-import { makeSideshiftPlugin } from './swap/sideshift'
-import { makeSwapuzPlugin } from './swap/swapuz'
 import { makeTransferPlugin } from './swap/transfer'
 
 const plugins = {
