+++ src/swap/central/godex.ts
@@ -1,5 +1,13 @@
-import { gt, lt } from 'biggystring'
-import { asObject, asOptional, asString } from 'cleaners'
+import { lt } from 'biggystring'
+import {
+  asArray,
+  asEither,
+  asMaybe,
+  asNull,
+  asObject,
+  asOptional,
+  asString
+} from 'cleaners'
 import {
   EdgeCorePluginOptions,
   EdgeSpendInfo,
@@ -7,7 +15,7 @@ import {
   EdgeSwapPlugin,
   EdgeSwapQuote,
   EdgeSwapRequest,
-  SwapAboveLimitError,
+  JsonObject,
   SwapBelowLimitError,
   SwapCurrencyError
 } from 'edge-core-js/types'
@@ -19,86 +27,87 @@ import {
   InvalidCurrencyCodes,
   makeSwapPluginQuote,
   SwapOrder
-} from '../util/swapHelpers'
-import { convertRequest, getAddress } from '../util/utils'
-import { asOptionalBlank } from './changenow'
-import { asNumberString, EdgeSwapRequestPlugin } from './types'
+} from '../../util/swapHelpers'
+import { convertRequest, getAddress } from '../../util/utils'
+import { asNumberString, EdgeSwapRequestPlugin } from '../types'
 
-const pluginId = 'letsexchange'
+const pluginId = 'godex'
 
 const swapInfo: EdgeSwapInfo = {
   pluginId,
   isDex: false,
-  displayName: 'LetsExchange',
-  supportEmail: 'support@letsexchange.io'
+  displayName: 'Godex',
+  supportEmail: 'support@godex.io'
 }
 
 const asInitOptions = asObject({
-  apiKey: asString,
-  affiliateId: asOptional(asString)
+  apiKey: asOptional(asString)
 })
 
-const orderUri = 'https://letsexchange.io/?exchangeId='
-const uri = 'https://api.letsexchange.io/api/v1/'
+const orderUri = 'https://godex.io/exchange/waiting/'
+const uri = 'https://api.godex.io/api/v1/'
 
 const expirationMs = 1000 * 60
 
+const asApiInfo = asObject({
+  min_amount: asNumberString,
+  networks_from: asMaybe(
+    asArray(
+      asObject({
+        network: asString
+      })
+    )
+  ),
+  networks_to: asMaybe(
+    asArray(
+      asObject({
+        network: asString
+      })
+    )
+  )
+})
+
 const asQuoteInfo = asObject({
   transaction_id: asString,
-  deposit_amount: asString,
   deposit: asString,
-  deposit_extra_id: asOptionalBlank(asString),
+  deposit_extra_id: asEither(asString, asNull),
+  deposit_amount: asString,
+  withdrawal: asString,
+  withdrawal_extra_id: asEither(asString, asNull),
   withdrawal_amount: asString,
-  withdrawal_extra_id: asOptionalBlank(asString)
+  return: asString,
+  return_extra_id: asEither(asString, asNull)
 })
 
-const asInfoReply = asObject({
-  min_amount: asNumberString,
-  max_amount: asNumberString,
-  amount: asNumberString
-})
 const INVALID_CURRENCY_CODES: InvalidCurrencyCodes = {
   from: {
-    ethereum: ['MATH'],
+    avalanche: 'allTokens',
+    celo: 'allTokens',
+    digibyte: 'allCodes',
+    ethereum: ['MATIC'],
+    fantom: 'allTokens',
     optimism: ['VELO'],
-    polygon: ['USDC.e']
+    polygon: 'allCodes'
   },
   to: {
-    ethereum: ['MATH'],
-    polygon: ['USDC.e'],
-    zcash: ['ZEC']
+    avalanche: 'allTokens',
+    celo: 'allTokens',
+    ethereum: ['MATIC'],
+    fantom: 'allTokens',
+    polygon: 'allCodes',
+    zcash: ['ZEC'] // ChangeHero doesn't support sending to unified addresses
   }
 }
 
-// See https://letsexchange.io/exchange-pairs for list of supported currencies
+// Network names that don't match parent network currency code
+// See https://godex.io/exchange-rate for list of supported currencies
 const MAINNET_CODE_TRANSCRIPTION = {
-  avalanche: 'AVAXC',
-  binance: 'BEP2',
-  binancesmartchain: 'BEP20',
-  ethereum: 'ERC20',
-  optimism: 'OPTIMISM',
   rsk: 'RSK',
-  tron: 'TRC20'
+  binancesmartchain: 'BSC',
+  avalanche: 'AVAXC'
 }
 
-const SPECIAL_MAINNET_CASES: { [pId: string]: { [cc: string]: string } } = {
-  avalanche: {
-    AVAX: 'AVAXC'
-  },
-  binancesmartchain: {
-    BNB: 'BEP20'
-  },
-  optimism: {
-    ETH: 'OPTIMISM'
-  },
-  zksync: {
-    ETH: 'ZKSYNC'
-  }
-}
-
-export function makeLetsExchangePlugin(
-  opts: EdgeCorePluginOptions
-): EdgeSwapPlugin {
+export function makeGodexPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
   const { io, log } = opts
   const { fetchCors = io.fetch } = io
   const initOptions = asInitOptions(opts.initOptions)
@@ -106,13 +115,12 @@ export function makeLetsExchangePlugin(
   async function call(
     url: string,
     request: EdgeSwapRequestPlugin,
-    data: { params: Object }
-  ): Promise<Object> {
+    data: { params: JsonObject }
+  ): Promise<JsonObject> {
     const body = JSON.stringify(data.params)
 
     const headers = {
       'Content-Type': 'application/json',
-      Authorization: `Bearer ${initOptions.apiKey}`,
       Accept: 'application/json'
     }
     const response = await fetchCors(url, { method: 'POST', body, headers })
@@ -120,7 +128,7 @@ export function makeLetsExchangePlugin(
       if (response.status === 422) {
         throw new SwapCurrencyError(swapInfo, request)
       }
-      throw new Error(`letsexchange returned error code ${response.status}`)
+      throw new Error(`godex returned error code ${response.status}`)
     }
     return await response.json()
   }
@@ -137,6 +145,11 @@ export function makeLetsExchangePlugin(
       getAddress(request.toWallet)
     ])
 
+    const { fromMainnetCode, toMainnetCode } = getCodesWithTranscription(
+      request,
+      MAINNET_CODE_TRANSCRIPTION
+    )
+
     // Convert the native amount to a denomination:
     const quoteAmount =
       request.quoteFor === 'from'
@@ -149,40 +162,15 @@ export function makeLetsExchangePlugin(
             request.toCurrencyCode
           )
 
-    const { fromMainnetCode, toMainnetCode } = getCodesWithTranscription(
-      request,
-      MAINNET_CODE_TRANSCRIPTION
-    )
-
-    const { pluginId: fromPluginId } = request.fromWallet.currencyInfo
-    const networkFrom =
-      SPECIAL_MAINNET_CASES[fromPluginId]?.[request.toCurrencyCode] != null
-        ? SPECIAL_MAINNET_CASES[fromPluginId][request.toCurrencyCode]
-        : request.fromCurrencyCode ===
-          request.fromWallet.currencyInfo.currencyCode
-        ? request.fromCurrencyCode
-        : fromMainnetCode
-
-    const { pluginId: toPluginId } = request.toWallet.currencyInfo
-    const networkTo =
-      SPECIAL_MAINNET_CASES[toPluginId]?.[request.toCurrencyCode] != null
-        ? SPECIAL_MAINNET_CASES[toPluginId][request.toCurrencyCode]
-        : request.toCurrencyCode === request.toWallet.currencyInfo.currencyCode
-        ? request.toCurrencyCode
-        : toMainnetCode
-
     // Swap the currencies if we need a reverse quote:
     const quoteParams = {
       from: request.fromCurrencyCode,
       to: request.toCurrencyCode,
-      network_from: networkFrom,
-      network_to: networkTo,
       amount: quoteAmount
     }
-
     log('quoteParams:', quoteParams)
 
-    // Calculate the amounts:
+    // Check if we are below the minimum limit:
     let fromAmount, toAmount, endpoint
     if (request.quoteFor === 'from') {
       fromAmount = quoteAmount
@@ -194,9 +182,11 @@ export function makeLetsExchangePlugin(
     const response = await call(uri + endpoint, request, {
       params: quoteParams
     })
-    const reply = asInfoReply(response)
+    const reply = asApiInfo(response)
 
-    // Check the min/max:
+    // The info/info-revert endpoints don't accept the coin_from/coin_to params so the
+    // min_amount returned could be for a different network than the user is requesting.
+    // Check the minimum:
     const nativeMin = reverseQuote
       ? await request.toWallet.denominationToNative(
           reply.min_amount,
@@ -215,51 +205,43 @@ export function makeLetsExchangePlugin(
       )
     }
 
-    const nativeMax = reverseQuote
-      ? await request.toWallet.denominationToNative(
-          reply.max_amount,
-          request.toCurrencyCode
-        )
-      : await request.fromWallet.denominationToNative(
-          reply.max_amount,
-          request.fromCurrencyCode
-        )
-
-    if (gt(nativeMax, '0')) {
-      if (gt(request.nativeAmount, nativeMax)) {
-        throw new SwapAboveLimitError(
-          swapInfo,
-          nativeMin,
-          reverseQuote ? 'to' : 'from'
-        )
-      }
+    // Check the networks. Networks aren't present for disabled assets.
+    if (
+      reply.networks_from?.find(
+        network => network.network === fromMainnetCode
+      ) == null ||
+      reply.networks_to?.find(network => network.network === toMainnetCode) ==
+        null
+    ) {
+      throw new SwapCurrencyError(swapInfo, request)
     }
 
     const { promoCode } = opts
+
     endpoint = reverseQuote ? 'transaction-revert' : 'transaction'
-    const sendReply = await call(uri + endpoint, request, {
-      params: {
-        deposit_amount: reverseQuote ? undefined : fromAmount,
-        withdrawal_amount: reverseQuote ? toAmount : undefined,
-        coin_from: request.fromCurrencyCode,
-        coin_to: request.toCurrencyCode,
-        network_from: networkFrom,
-        network_to: networkTo,
-        withdrawal: toAddress,
-        return: fromAddress,
-        return_extra_id: null,
-        withdrawal_extra_id: null,
-        affiliate_id: initOptions.affiliateId,
-        promocode: promoCode != null ? promoCode : '',
-        type: 'edge',
-        float: false,
-        isEstimate: false
+    const sendReply = await call(
+      uri + endpoint + (promoCode != null ? `?promo=${promoCode}` : ''),
+      request,
+      {
+        params: {
+          deposit_amount: reverseQuote ? undefined : fromAmount,
+          withdrawal_amount: reverseQuote ? toAmount : undefined,
+          coin_from: request.fromCurrencyCode,
+          coin_to: request.toCurrencyCode,
+          withdrawal: toAddress,
+          return: fromAddress,
+          return_extra_id: null,
+          withdrawal_extra_id: null,
+          affiliate_id: initOptions.apiKey,
+          type: 'edge',
+          isEstimate: false,
+          coin_from_network: fromMainnetCode,
+          coin_to_network: toMainnetCode
+        }
       }
-    })
-
-    log('sendReply', sendReply)
+    )
+    log('sendReply' + JSON.stringify(sendReply, null, 2))
     const quoteInfo = asQuoteInfo(sendReply)
-
     const fromNativeAmount = await request.fromWallet.denominationToNative(
       quoteInfo.deposit_amount,
       request.fromCurrencyCode
@@ -269,6 +251,9 @@ export function makeLetsExchangePlugin(
       request.toCurrencyCode
     )
 
+    log('fromNativeAmount: ' + fromNativeAmount)
+    log('toNativeAmount: ' + toNativeAmount)
+
     // Make the transaction:
     const spendInfo: EdgeSpendInfo = {
       tokenId: request.fromTokenId,
@@ -276,7 +261,7 @@ export function makeLetsExchangePlugin(
         {
           nativeAmount: fromNativeAmount,
           publicAddress: quoteInfo.deposit,
-          uniqueIdentifier: quoteInfo.deposit_extra_id
+          uniqueIdentifier: quoteInfo.deposit_extra_id ?? undefined
         }
       ],
       networkFeeOption:
@@ -305,7 +290,6 @@ export function makeLetsExchangePlugin(
         refundAddress: fromAddress
       }
     }
-
     log('spendInfo', spendInfo)
 
     return {
